refactor(Aliases): make createAlisesModal his own component
